% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yy_plt_forest_lm.R
\name{yy_plt_forest_binom}
\alias{yy_plt_forest_binom}
\title{Make a plot from an lm item}
\usage{
yy_plt_forest_binom(binom_model)
}
\arguments{
\item{binom_model}{Binomial model from base R}
}
\value{
Forest plot, based on ggplot2
}
\description{
Make a plot from an lm item
}
