% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yy_plt_forest_lm.R
\name{yy_plt_forest_lm}
\alias{yy_plt_forest_lm}
\title{Make a plot from an glm, binomal item}
\usage{
yy_plt_forest_lm(lm_model)
}
\arguments{
\item{lm_model}{Linear model from base R}
}
\value{
Forest plot, based on ggplot2
}
\description{
Make a plot from an glm, binomal item
}
